Minor edits
Created CSS Class: "responsive-button" that has been applied to the nav-bar buttons on each page as well as the main menu buttons on menu/activities.  Has the property of on-hover increasing opacity by 10% and transforming the size 30%.

ID:menu-button additionally applies a contrast filter of 110% to differentiate colors within the user-profile screen.